Really fix "X needs update" bug

+++ make.sh
@@ -186,11 +186,11 @@ fi
 
 if [ -f Makefile ]; then
   # FIXME check wont't work when PROJECT is a multi-dir path
-  if [ "$PROJECT" = Gaudi ]; then
+  if [ "$PROJECT" = Gaudi ] && [ -f "$PROJECT/Makefile" ]; then
     # FIXME remove the following once Makefile is removed from Gaudi
     # update-index is needed as older git diff-index only does a "quick look"
     # see https://stackoverflow.com/a/34808299
-    gitc "$PROJECT" update-index --refresh 2>/dev/null || true
+    gitc "$PROJECT" update-index --refresh &>/dev/null || true
     if gitc "$PROJECT" diff-index --quiet HEAD -- Makefile; then
       rm -f "$PROJECT/Makefile"
       # hide the removed file from status and diffs
